requirements.txt
quart==0.18.4
aiohttp==3.8.5
discord.py==2.3.2